Update index.html
The healing map takes the viewer on a journey in the recreation ground of Kivinokka where one can explore nature with simple methods that extend their stay in the area. With these simple methods one can maximize the consumption of Mycobacterium vaccae (a soil bacterium) and enjoy the release of serotonin and noradrenaline in the brain by simply breathing in nature.